+++ common/tools/dev-tool/src/util/customization/customize.ts
@@ -202,11 +202,7 @@ export function mergeModuleDeclarations(
     originalVirtualSourceFile
   );
 
-  augmentImports(
-    originalDeclarationsMap.imports,
-    customVirtualSourceFile.getImportDeclarations(),
-    originalVirtualSourceFile
-  );
+  augmentImports(originalDeclarationsMap.imports, customVirtualSourceFile.getImportDeclarations());
 
   originalVirtualSourceFile.fixMissingImports();
   sortSourceFileContents(originalVirtualSourceFile);
